scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "XuChenXu", email = "91937041+ChenXu233@users.noreply.github.com"},
]
dependencies = [
    "fastapi>=0.124.2",
    "uvicorn>=0.38.0",
    "cryptography>=46.0.3",
    "aiohttp>=3.13.2",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
"dev" = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
"start" = "uvicorn app.main:app --host 0.0.0.0 --port 8000"
